# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Royal Dental Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medical-records:
    get:
      operationId: listMedicalRecords
      tags: [medical-records]
      summary: List medical records for a patient
      parameters:
        - name: patientId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalRecord"
    post:
      operationId: createMedicalRecord
      tags: [medical-records]
      summary: Create medical record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicalRecordBody"
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medical-records/{id}:
    patch:
      operationId: updateMedicalRecord
      tags: [medical-records]
      summary: Update medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicalRecordBody"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List medications
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [medications]
      summary: Add medication (doctor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicationBody"
      responses:
        "201":
          description: Medication added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /medications/{id}:
    delete:
      operationId: deleteMedication
      tags: [medications]
      summary: Delete medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Medication deleted
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List conversations for current user
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [messages]
      summary: Start a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [dashboard]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  schemas:
...
    MedicalRecord:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        doctorName:
          type: string
        diagnosis:
          type: string
        treatment:
          type: string
        notes:
          type: ["string", "null"]
        toothNumber:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - doctorId
        - doctorName
        - diagnosis
        - treatment
        - createdAt
    CreateMedicalRecordBody:
      type: object
      properties:
        patientId:
          type: integer
        diagnosis:
          type: string
        treatment:
          type: string
        notes:
          type: string
        toothNumber:
          type: integer
      required:
        - patientId
        - diagnosis
        - treatment
    UpdateMedicalRecordBody:
      type: object
      properties:
        diagnosis:
          type: string
        treatment:
          type: string
        notes:
          type: string
        toothNumber:
          type: integer
    Medication:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        dosageForm:
          type: string
        strength:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - dosageForm
        - strength
        - createdAt
    CreateMedicationBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        dosageForm:
          type: string
        strength:
          type: string
        description:
          type: string
      required:
        - name
        - category
        - dosageForm
        - strength
...
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - senderId
        - senderName
        - content
        - createdAt
    SendMessageBody:
      type: object
      properties:
        conversationId:
          type: integer
        content:
          type: string
      required:
        - conversationId
        - content
    Conversation:
      type: object
      properties:
        id:
          type: integer
        participantIds:
          type: array
          items:
            type: integer
        participantNames:
          type: array
          items:
            type: string
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - participantIds
        - participantNames
        - unreadCount
        - createdAt
    CreateConversationBody:
      type: object
      properties:
        participantId:
          type: integer
      required:
        - participantId
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [appointment, prescription, system, reminder]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt